+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased] - XXXX-XX-XX
 
+### Added
+- Add increment option to fix-encrypted-version command [#279](https://github.com/owncloud/encryption/issues/279)
+
 ### Fixed
 
 - Determine encryption format correctly when using HSM  [#261](https://github.com/owncloud/encryption/pull/261)